Added a first try on an emulateTransitionEnd function.
After some profiling, it turns out that the whole transitionEnd listening is by far the most expensive part of the modal dialog.
Since the worst case scenary for just using a simple setTimeout method is that the modal + overlay is turned off before the transition is done,
and that really means no harm, I am just skipping all the transitionEnd stuff, so this commit is solely for keeping all code in the repos.
In other words: This method isn't done, and it never will be.